refactor: fuse unicast and multicast tx
