+++ include/zenoh-pico/session/utils.h
@@ -31,9 +31,6 @@ void _z_session_clear(_z_session_t *zn);
 z_result_t _z_session_close(_z_session_t *zn, uint8_t reason);
 
 z_result_t _z_handle_network_message(_z_session_rc_t *zsrc, _z_zenoh_message_t *z_msg, uint16_t local_peer_id);
-z_result_t _z_send_n_msg(_z_session_t *zn, _z_network_message_t *n_msg, z_reliability_t reliability,
-                         z_congestion_control_t cong_ctrl);
-z_result_t _z_send_n_batch(_z_session_t *zn, z_congestion_control_t cong_ctrl);
 
 #if Z_FEATURE_MULTI_THREAD == 1
 static inline void _z_session_mutex_lock(_z_session_t *zn) { (void)_z_mutex_lock(&zn->_mutex_inner); }
