+++ src/net/query.c
@@ -14,6 +14,7 @@
 #include "zenoh-pico/net/query.h"
 
 #include "zenoh-pico/session/utils.h"
+#include "zenoh-pico/transport/common/tx.h"
 #include "zenoh-pico/utils/logging.h"
 
 void _z_query_clear_inner(_z_query_t *q) {
